+++ src/components/input/template.test.js
@@ -73,6 +73,17 @@ describe('Input', () => {
       expect($component.val()).toEqual('QQ 12 34 56 C')
     })
 
+    it('renders with aria-describedby', () => {
+      const describedById = 'some-id'
+
+      const $ = render('input', {
+        describedBy: describedById
+      })
+
+      const $component = $('.govuk-input')
+      expect($component.attr('aria-describedby')).toMatch(describedById)
+    })
+
     it('renders with attributes', () => {
       const $ = render('input', {
         attributes: {
@@ -133,6 +144,28 @@ describe('Input', () => {
       expect($input.attr('aria-describedby'))
         .toMatch(hintId)
     })
+
+    it('associates the input as "described by" the hint and parent fieldset', () => {
+      const describedById = 'some-id'
+
+      const $ = render('input', {
+        id: 'input-with-hint',
+        describedBy: describedById,
+        hint: {
+          text: 'It’s on your National Insurance card, benefit letter, payslip or P60. For example, ‘QQ 12 34 56 C’.'
+        }
+      })
+
+      const $input = $('.govuk-input')
+      const $hint = $('.govuk-hint')
+
+      const hintId = new RegExp(
+        WORD_BOUNDARY + describedById + WHITESPACE + $hint.attr('id') + WORD_BOUNDARY
+      )
+
+      expect($input.attr('aria-describedby'))
+        .toMatch(hintId)
+    })
   })
 
   describe('when it includes an error message', () => {
@@ -160,6 +193,28 @@ describe('Input', () => {
 
       const errorMessageId = new RegExp(
         WORD_BOUNDARY + $errorMessage.attr('id') + WORD_BOUNDARY
+      )
+
+      expect($input.attr('aria-describedby'))
+        .toMatch(errorMessageId)
+    })
+
+    it('associates the input as "described by" the error message and parent fieldset', () => {
+      const describedById = 'some-id'
+
+      const $ = render('input', {
+        id: 'input-with-error',
+        describedBy: describedById,
+        errorMessage: {
+          text: 'Error message'
+        }
+      })
+
+      const $input = $('.govuk-input')
+      const $errorMessage = $('.govuk-error-message')
+
+      const errorMessageId = new RegExp(
+        WORD_BOUNDARY + describedById + WHITESPACE + $errorMessage.attr('id') + WORD_BOUNDARY
       )
 
       expect($input.attr('aria-describedby'))
@@ -201,11 +256,36 @@ describe('Input', () => {
       })
 
       const $component = $('.govuk-input')
-      const $errorMessageId = $('.govuk-error-message').attr('id')
-      const $hintId = $('.govuk-hint').attr('id')
+      const errorMessageId = $('.govuk-error-message').attr('id')
+      const hintId = $('.govuk-hint').attr('id')
 
       const combinedIds = new RegExp(
-        WORD_BOUNDARY + $hintId + WHITESPACE + $errorMessageId + WORD_BOUNDARY
+        WORD_BOUNDARY + hintId + WHITESPACE + errorMessageId + WORD_BOUNDARY
+      )
+
+      expect($component.attr('aria-describedby'))
+        .toMatch(combinedIds)
+    })
+
+    it('associates the input as described by the hint, error message and parent fieldset', () => {
+      const describedById = 'some-id'
+
+      const $ = render('input', {
+        describedBy: describedById,
+        errorMessage: {
+          text: 'Error message'
+        },
+        hint: {
+          text: 'Hint'
+        }
+      })
+
+      const $component = $('.govuk-input')
+      const errorMessageId = $('.govuk-error-message').attr('id')
+      const hintId = $('.govuk-hint').attr('id')
+
+      const combinedIds = new RegExp(
+        WORD_BOUNDARY + describedById + WHITESPACE + hintId + WHITESPACE + errorMessageId + WORD_BOUNDARY
       )
 
       expect($component.attr('aria-describedby'))