+++ src/components/select/template.test.js
@@ -135,6 +135,17 @@ describe('Select', () => {
       expect($firstItem.attr('disabled')).toBeTruthy()
     })
 
+    it('renders with aria-describedby', () => {
+      const describedById = 'some-id'
+
+      const $ = render('select', {
+        describedBy: describedById
+      })
+
+      const $component = $('.govuk-select')
+      expect($component.attr('aria-describedby')).toMatch(describedById)
+    })
+
     it('renders with attributes', () => {
       const $ = render('select', {
         attributes: {
@@ -229,6 +240,28 @@ describe('Select', () => {
       expect($select.attr('aria-describedby'))
         .toMatch(hintId)
     })
+
+    it('associates the select as "described by" the hint and parent fieldset', () => {
+      const describedById = 'some-id'
+
+      const $ = render('select', {
+        id: 'select-with-hint',
+        describedBy: describedById,
+        hint: {
+          'text': 'Hint text goes here'
+        }
+      })
+
+      const $select = $('.govuk-select')
+      const $hint = $('.govuk-hint')
+
+      const hintId = new RegExp(
+        WORD_BOUNDARY + describedById + WHITESPACE + $hint.attr('id') + WORD_BOUNDARY
+      )
+
+      expect($select.attr('aria-describedby'))
+        .toMatch(hintId)
+    })
   })
 
   describe('when it includes an error message', () => {
@@ -256,6 +289,28 @@ describe('Select', () => {
 
       const errorMessageId = new RegExp(
         WORD_BOUNDARY + $errorMessage.attr('id') + WORD_BOUNDARY
+      )
+
+      expect($input.attr('aria-describedby'))
+        .toMatch(errorMessageId)
+    })
+
+    it('associates the select as "described by" the error message and parent fieldset', () => {
+      const describedById = 'some-id'
+
+      const $ = render('select', {
+        id: 'select-with-error',
+        describedBy: describedById,
+        errorMessage: {
+          'text': 'Error message'
+        }
+      })
+
+      const $input = $('.govuk-select')
+      const $errorMessage = $('.govuk-error-message')
+
+      const errorMessageId = new RegExp(
+        WORD_BOUNDARY + describedById + WHITESPACE + $errorMessage.attr('id') + WORD_BOUNDARY
       )
 
       expect($input.attr('aria-describedby'))
@@ -297,11 +352,36 @@ describe('Select', () => {
       })
 
       const $component = $('.govuk-select')
-      const $errorMessageId = $('.govuk-error-message').attr('id')
-      const $hintId = $('.govuk-hint').attr('id')
+      const errorMessageId = $('.govuk-error-message').attr('id')
+      const hintId = $('.govuk-hint').attr('id')
 
       const combinedIds = new RegExp(
-        WORD_BOUNDARY + $hintId + WHITESPACE + $errorMessageId + WORD_BOUNDARY
+        WORD_BOUNDARY + hintId + WHITESPACE + errorMessageId + WORD_BOUNDARY
+      )
+
+      expect($component.attr('aria-describedby'))
+        .toMatch(combinedIds)
+    })
+
+    it('associates the select as described by the hint, error message and parent fieldset', () => {
+      const describedById = 'some-id'
+
+      const $ = render('select', {
+        describedBy: describedById,
+        errorMessage: {
+          text: 'Error message'
+        },
+        hint: {
+          text: 'Hint'
+        }
+      })
+
+      const $component = $('.govuk-select')
+      const errorMessageId = $('.govuk-error-message').attr('id')
+      const hintId = $('.govuk-hint').attr('id')
+
+      const combinedIds = new RegExp(
+        WORD_BOUNDARY + describedById + WHITESPACE + hintId + WHITESPACE + errorMessageId + WORD_BOUNDARY
       )
 
       expect($component.attr('aria-describedby'))